+++ com.saab.map-streamer/Runtime/Saab.Foundation.Unity/Saab.Foundation.Unity.MapStreamer/GeodeticCamera/CameraShaderUpdater.cs
@@ -35,7 +35,7 @@ using Saab.Unity.Extensions;
 using System;
 using UnityEngine;
 
-namespace Saab.Foundation.Unity.MapStreamer
+namespace Saab.Foundation.Unity.MapStreamer.GeodeticCamera
 {
     public class CameraShaderUpdater : MonoBehaviour
     {